+++ packages/tools/doc-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @larkup/tool-doc-editor
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @larkup/core@0.2.3
+
 ## 0.2.3
 
 ### Patch Changes